requirements.txt
together
faiss-cpu
click
rich
pyyaml
python-dotenv
numpy
gitpython
requests
aiohttp 